Implement Newsletter and Sign-up.
Implement a working Sign-Up page which adds the email to a mongo-db
database on the server. The server sends a mail using nodemailer and
gmail.

The server sends message to all the emails in the database every week.
It reads what message is to be send from a html file on the disk.